+++ tests/core_tests/chaingen_main.cpp
@@ -122,6 +122,7 @@ int main(int argc, char* argv[])
     GENERATE_AND_PLAY(gen_block_ts_at_median);
     GENERATE_AND_PLAY(gen_block_alt_ts_above_ftl);
     GENERATE_AND_PLAY(gen_block_alt_ts_window_truncation);
+    GENERATE_AND_PLAY(gen_reorg_watermark_refused_switch);
     GENERATE_AND_PLAY(gen_block_ts_in_future);
     GENERATE_AND_PLAY(gen_block_invalid_prev_id);
     GENERATE_AND_PLAY(gen_block_invalid_attestation_root);